+++ docs/src/ngdoc.js
@@ -726,7 +726,7 @@ function scenarios(docs){
 
 //////////////////////////////////////////////////////////
 function metadata(docs){
-  var words = [];
+  var pages = [];
   docs.forEach(function(doc){
     var path = (doc.name || '').split(/(\.|\:\s+)/);
     for ( var i = 1; i < path.length; i++) {
@@ -738,7 +738,7 @@ function metadata(docs){
       shortName = 'input [' + shortName + ']';
     }
 
-    words.push({
+    pages.push({
       section: doc.section,
       id: doc.id,
       name: title(doc.name),
@@ -747,18 +747,22 @@ function metadata(docs){
       keywords:doc.keywords()
     });
   });
-  words.sort(keywordSort);
-  return words;
+  pages.sort(keywordSort);
+  return pages;
 }
 
 var KEYWORD_PRIORITY = {
   '.index': 1,
-  '.guide': 2,
-  '.angular': 7,
-  '.angular.Module': 7,
-  '.angular.module': 8,
-  '.angular.module.ng': 2,
-  '.angular.module.AUTO': 1,
+  '.overview': 1,
+  '.bootstrap': 2,
+  '.mvc': 3,
+  '.scopes': 4,
+  '.compiler': 5,
+  '.templates': 6,
+  '.services': 7,
+  '.di': 8,
+  '.unit-testing': 9,
+  '.dev_guide': 9,
   '.dev_guide.overview': 1,
   '.dev_guide.bootstrap': 2,
   '.dev_guide.bootstrap.auto_bootstrap': 1,